% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{add_significance_flag}
\alias{add_significance_flag}
\title{Add significance flag}
\usage{
add_significance_flag(data, p_val_col)
}
\arguments{
\item{data}{data frame containing a column of p-values}

\item{p_val_col}{name of the p-value column}
}
\value{
a data frame with an additional character column named p_flag
}
\description{
Add significance flag
}
