% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{as_csv_string}
\alias{as_csv_string}
\title{Convert a data frame to a csv string}
\usage{
as_csv_string(x)
}
\arguments{
\item{x}{the object to be written, preferably a matrix or data frame. If not,
it is attempted to coerce x to a data frame.}
}
\value{
a csv character string representation of x
}
\description{
Convert a data frame to a csv string
}
\examples{
as_csv_string(mtcars)
}
