% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{assign_in_global}
\alias{assign_in_global}
\title{Assign a Value to a Name}
\usage{
assign_in_global(value, x, print = FALSE, envir = .GlobalEnv, ...)
}
\arguments{
\item{value}{a value to be assigned to \code{x}.}

\item{x}{a variable name, given as a character string.  No coercion is
    done, and the first element of a character vector of length greater
    than one will be used, with a warning.}

\item{print}{logical whether to also print}

\item{envir}{the \code{\link[base]{environment}} to use.  See \sQuote{Details}.}

\item{...}{additional arguments passed to \link[base]{assign}}
}
\value{
value object (invisibly)
}
\description{
Assign a Value to a Name
}
\examples{
## print and assign copy of mtcars
mtcars \%>\% print() \%>\% assign_in_global("mtcars2")

## same as above but using print argument
mtcars \%>\% assign_in_global("mtcars2", TRUE)
}
\seealso{
\link[base]{assign}
}
