% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{bib_add_pkg}
\alias{bib_add_pkg}
\title{Add R package to bibliography}
\usage{
bib_add_pkg(package = "base", bibliography = "bib/bib.bib", ...)
}
\arguments{
\item{package}{a character string with the name of a single package.
    An error occurs if more than one package name is given.}

\item{bibliography}{path to bibliography file}

\item{...}{additional parameters passed to \code{\link[utils]{citation}}}
}
\description{
Add R package to bibliography
}
