% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bib_bibliography}
\alias{bib_bibliography}
\title{Determine bibliography from RMarkdown YAML}
\usage{
bib_bibliography(dirname = FALSE)
}
\arguments{
\item{dirname}{flag to return the bibliography directory instead of the full relative path}
}
\value{
full bibliography path specified in the RMarkdown YAML or parent directory
}
\description{
Determine bibliography from RMarkdown YAML
}
