% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bib_project_collection}
\alias{bib_project_collection}
\title{Determine Zotero collection name from open RMarkdown file}
\usage{
bib_project_collection()
}
\value{
name of Zotero collection
}
\description{
Determine Zotero collection name from open RMarkdown file
}
