% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bib_sync_zotero}
\alias{bib_sync_zotero}
\title{Download Zotero collection}
\usage{
bib_sync_zotero(
  download_pdfs = TRUE,
  collection_name = bib_zotero_collection(),
  bibliography = bib_bibliography(),
  ...
)
}
\arguments{
\item{download_pdfs}{flag to download pdfs (if any)}

\item{collection_name}{name of the collection to download from}

\item{bibliography}{path to the bibliography file}

\item{...}{additional parameters passed to httr-based functions}
}
\description{
Download Zotero collection
}
