% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{bib_use_csl}
\alias{bib_use_csl}
\title{Use citation style language}
\usage{
bib_use_csl(csl = c("american-medical-association", "apa"), bib_path = "bib")
}
\arguments{
\item{csl}{citation style language}

\item{bib_path}{path to bibliography directory}
}
\description{
Use citation style language
}
\references{
\href{https://github.com/citation-style-language/styles}{Citation Style Language Styles}
}
