% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R
\name{bib_zotero_collection}
\alias{bib_zotero_collection}
\title{Determine Zotero collection name from open RMarkdown file name}
\usage{
bib_zotero_collection()
}
\value{
name of Zotero collection
}
\description{
Determine Zotero collection name from open RMarkdown file name
}
