% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy.glmerMod}
\alias{broom_tidy.glmerMod}
\title{Tidy a glmerMod object}
\usage{
broom_tidy.glmerMod(x, ...)
}
\arguments{
\item{x}{A glmerMod object returned from \code{\link[lme4:glmer]{glmer()}}}

\item{...}{Additional arguments. Not used. Needed to match generic signature
only. Cautionary note: Misspelled arguments will be absorbed in ..., where
they will be ignored. If the misspelled argument has a default value, the
default value will be used.}
}
\value{
A \link[tibble]{tibble} with information about fixed model estimates.
}
\description{
Tidy a glmerMod object
}
\examples{
m <- lme4::glmer(vs ~ mpg + (1|cyl), mtcars, binomial)
broom_tidy.glmerMod(m)
}
