% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy_cor_test}
\alias{broom_tidy_cor_test}
\title{Pairwise cor tests}
\usage{
broom_tidy_cor_test(data, ...)
}
\arguments{
\item{data}{data frame containing the variables}

\item{...}{further arguments passed to \code{\link[stats:cor.test]{cor.test()}}}
}
\value{
a tibble with information about the model components; one model per row
}
\description{
Test for association between paired samples, using one of Pearson's product moment correlation coefficient, Kendall's tau or Spearman's rho for each pair of variables.
}
