% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy_pairwise_odds_ratio}
\alias{broom_tidy_pairwise_odds_ratio}
\title{Pairwise odds ratios}
\usage{
broom_tidy_pairwise_odds_ratio(
  data,
  response,
  exposure,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response and exposure variables}

\item{response}{response variable}

\item{exposure}{exposure variable}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{...}{additional parameters ultimately passed onto \code{\link[epitools:oddsratio]{oddsratio()}}}
}
\value{
a tibble with computed odds ratios and associated information
}
\description{
Computes odds ratios for each pair of exposures from tidy data.
}
