% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy_pairwise_t_test.R
\name{broom_tidy_pairwise_t_test_dep_sample}
\alias{broom_tidy_pairwise_t_test_dep_sample}
\title{Pairwise t tests}
\usage{
broom_tidy_pairwise_t_test_dep_sample(data, response, group, unit, ...)
}
\arguments{
\item{data}{tidy data containing the response, group, and unit variables}

\item{response}{response variable}

\item{group}{grouping variable}

\item{unit}{sampling unit variable (e.g., persons)}

\item{...}{additional parameters passed onto t.test()}
}
\value{
a tibble with information about the model components; one model per row
}
\description{
Pairwise t tests
}
