% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy_pairwise_t_test.R
\name{broom_tidy_pairwise_t_test_one_sample}
\alias{broom_tidy_pairwise_t_test_one_sample}
\title{Pairwise t tests}
\usage{
broom_tidy_pairwise_t_test_one_sample(data, response, group, mu = 0, ...)
}
\arguments{
\item{data}{tidy data containing the response and group variables}

\item{response}{response variable}

\item{group}{grouping variable}

\item{mu}{a number indicating the true value of the mean}

\item{...}{additional parameters passed onto t.test()}
}
\value{
a tibble with information about the model components; one model per row
}
\description{
Pairwise t tests
}
