% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom_tidy.R
\name{broom_tidy_pairwise_t_test_two_sample}
\alias{broom_tidy_pairwise_t_test_two_sample}
\title{Pairwise t tests}
\usage{
broom_tidy_pairwise_t_test_two_sample(data, response, group, ...)
}
\arguments{
\item{data}{tidy data containing the response and group variables}

\item{response}{response variable}

\item{group}{grouping variable}

\item{...}{additional parameters passed onto \code{\link[stats:t.test]{t.test()}}}
}
\value{
a tibble with information about the model components; one model per row
}
\description{
Performs independent sample t-tests for each pair of groups from tidy data.
}
