% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy_pairwise_wilcox_test_two_sample}
\alias{broom_tidy_pairwise_wilcox_test_two_sample}
\title{Pairwise Wilcoxon (Mann Whitney U) Rank Sum Tests}
\usage{
broom_tidy_pairwise_wilcox_test_two_sample(
  data,
  response,
  group,
  descriptives = TRUE,
  ...
)
}
\arguments{
\item{data}{tidy data containing the response and group variables}

\item{response}{response variable}

\item{group}{grouping variable}

\item{descriptives}{= logical to add descriptive statistics or not}

\item{...}{additional parameters passed onto \code{\link[stats:wilcox.test]{wilcox.test()}}}
}
\value{
a tibble with information about the model components; one model per row
}
\description{
Performs independent sample Wilcoxon tests for each pair of groups from tidy data.
}
