% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy_sided_ci}
\alias{broom_tidy_sided_ci}
\title{One-sided Confidence Intervals}
\usage{
broom_tidy_sided_ci(
  x,
  alternatives = "two.sided",
  conf.level = 0.95,
  exponentiate = FALSE
)
}
\arguments{
\item{x}{An lm object created by stats::lm().}

\item{alternatives}{Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the the coefficient estimates. This is typical for logistic and multinomial regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.}
}
\description{
One-sided Confidence Intervals
}
