% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_consult_exists}
\alias{cdb_consult_exists}
\title{Consult Database Check if Consult Exists}
\usage{
cdb_consult_exists(k)
}
\arguments{
\item{k}{consult id}
}
\value{
logical indicating whether the given consult already exists in \code{\link{cdb_file}}
}
\description{
Consult Database Check if Consult Exists
}
