% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_current_consult}
\alias{cdb_current_consult}
\title{Consult Database Current Consult}
\usage{
cdb_current_consult()
}
\value{
the consult id derived from the open source editor file name
}
\description{
Consult Database Current Consult
}
