% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_edit_app}
\alias{cdb_edit_app}
\title{Consult Database Edit App}
\usage{
cdb_edit_app(k = cdb_current_consult(), new = FALSE)
}
\arguments{
\item{k}{consult id}

\item{new}{logical; is this a new consult or should the consult data be read
from \code{\link{cdb_file}}?}
}
\description{
Consult Database Edit App
}
