% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_get_principal_investigator}
\alias{cdb_get_principal_investigator}
\title{Consult Database Get Principal Investigator}
\usage{
cdb_get_principal_investigator(k)
}
\arguments{
\item{k}{consult id}
}
\value{
tibble read from \code{\link{cdb_file}} representing the principal
investigator for the given consult
}
\description{
Consult Database Get Principal Investigator
}
