% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_open_current_consult_directory}
\alias{cdb_open_current_consult_directory}
\title{Open the Current Consult Directory}
\usage{
cdb_open_current_consult_directory()
}
\description{
Open the Current Consult Directory
}
