% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_sheet_upsert}
\alias{cdb_sheet_upsert}
\title{Consult Database Upsert}
\usage{
cdb_sheet_upsert(x, sheetName = c("consult", "personnel", "consult_personnel"))
}
\arguments{
\item{x}{data frame or data frame extension representing rows to insert or
update}

\item{sheetName}{a character string with the sheet name}
}
\description{
Consult Database Upsert
}
\details{
inserts or updates the database sheet depending on whether or not
the key value in x already exists in the sheet
}
