% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_template_consult}
\alias{cdb_template_consult}
\title{Consult Database New Consult Data Frame}
\usage{
cdb_template_consult(k)
}
\arguments{
\item{k}{consult id}
}
\value{
an "empty" consult data.frame
}
\description{
Consult Database New Consult Data Frame
}
