% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_template_consult_column_labels}
\alias{cdb_template_consult_column_labels}
\title{Consult Database New Consult Data Frame Column Labels}
\usage{
cdb_template_consult_column_labels()
}
\value{
character vector representing column labels for
\code{\link{cdb_template_consult}}
}
\description{
Consult Database New Consult Data Frame Column Labels
}
