% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_template_install}
\alias{cdb_template_install}
\title{Consult Database Install from Template}
\usage{
cdb_template_install(overwrite = FALSE)
}
\arguments{
\item{overwrite}{logical; should existing \code{\link{cdb_file}} be
overwritten?}
}
\value{
logical indicating if the copy succeeded
}
\description{
Consult Database Install from Template
}
