% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_template_personnel}
\alias{cdb_template_personnel}
\title{Consult Database New Personnel Data Frame}
\usage{
cdb_template_personnel()
}
\value{
an "empty" personnel data.frame
}
\description{
Consult Database New Personnel Data Frame
}
