% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{cdb_template_personnel_column_labels}
\alias{cdb_template_personnel_column_labels}
\title{Consult Database New Personnel Data Frame Column Labels}
\usage{
cdb_template_personnel_column_labels()
}
\value{
character vector representing column labels for \code{\link{cdb_template_personnel}}
}
\description{
Consult Database New Personnel Data Frame Column Labels
}
