% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check}
\alias{check}
\title{Build and check a package, cleaning up automatically on success.}
\usage{
check(document = FALSE, env_vars = c(`_R_CHECK_SYSTEM_CLOCK_` = FALSE), ...)
}
\arguments{
\item{document}{If \code{NA} and the package uses roxygen2, will
rerun \code{\link[devtools:document]{document()}} prior to checking. Use \code{TRUE}
and \code{FALSE} to override this default.}

\item{env_vars}{Environment variables set during \verb{R CMD check}}

\item{...}{additional parameters passed to \code{\link[devtools]{check}}}
}
\description{
Build and check a package, cleaning up automatically on success.
}
\note{
Adds environment flag to skip time check as per
\href{https://stat.ethz.ch/pipermail/r-package-devel/2019q1/003577.html}{r-package-devel}.
}
