% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{coef_convert_hr_to_beta}
\alias{coef_convert_hr_to_beta}
\title{Convert hazard ratio to beta coefficient}
\usage{
coef_convert_hr_to_beta(hr, se)
}
\arguments{
\item{hr}{hazard ratio}

\item{se}{standard error of the hazard ratio}
}
\value{
named list containing estimates for the beta coefficient and
associated standard error
}
\description{
Convert hazard ratio to beta coefficient
}
\examples{
coef_convert_hr_to_beta(1.3, .01)
}
