% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{consult_db_copy_user}
\alias{consult_db_copy_user}
\title{Copy user information from ldap to clipboard}
\usage{
consult_db_copy_user(email)
}
\arguments{
\item{email}{mail of the user}
}
\description{
Look up a user in ldap by email and copy data to clipboard to
paste into the consult database.
}
