% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{consult_report_document}
\alias{consult_report_document}
\title{Consult report document
Format for converting from R Markdown to a consult report document.}
\usage{
consult_report_document(
  toc = TRUE,
  toc_float = TRUE,
  toc_depth = 3,
  css,
  includes,
  pandoc_args,
  ...
)
}
\arguments{
\item{toc}{\code{TRUE} to include a table of contents in the output}

\item{toc_float}{\code{TRUE} to float the table of contents to the left of the
main document content. Rather than \code{TRUE} you may also pass a list of
options that control the behavior of the floating table of contents. See the
\emph{Floating Table of Contents} section below for details.}

\item{toc_depth}{Depth of headers to include in table of contents}

\item{css}{One or more css files to include}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{pandoc_args}{Additional command line options to pass to pandoc}

\item{...}{additional arguments passed to \code{\link[rmarkdown]{html_document}}}
}
\value{
R Markdown output format to pass to \code{\link[rmarkdown]{render}}
}
\description{
Consult report document
Format for converting from R Markdown to a consult report document.
}
\details{
Requires Pandoc version 2.10 or higher for use of --no-check-certificate option.
}
\seealso{
\url{https://pandoc.org/installing.html}
}
