% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contr.R
\name{contr.backward_difference}
\alias{contr.backward_difference}
\title{Backward difference contrast matrix}
\usage{
contr.backward_difference(fct)
}
\arguments{
\item{fct}{factor variable to apply contrast to}
}
\value{
matrix with n rows and n-1 columns where n is the number of factor
levels
}
\description{
Return a backward difference contrast matrix
}
\examples{
contrasts(iris$Species) <- contr.backward_difference(iris$Species)
}
\references{
\href{https://stats.oarc.ucla.edu/r/library/r-library-contrast-coding-systems-for-categorical-variables/#backward}{Backward difference coding} # nolint
}
\seealso{
\link[stats]{contrasts}
}
