% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cseq}
\alias{cseq}
\title{Centered sequence generation}
\usage{
cseq(center, by, length.out = 3)
}
\arguments{
\item{center}{the middle value of the sequence}

\item{by}{number: decrement/increment of the sequence}

\item{length.out}{desired length of the sequence}
}
\value{
vector of type integer or double
}
\description{
Generate regular sequences based on a central value
}
\details{
Generates a sequence of numbers above and below a central value.

The length.out argument should be an odd integer.
If the value given is fractional, it will be rounded up.
If the value given is even, it will be made odd by adding one.
}
\examples{
# integer center, integer by, odd length.out
cseq(center = 10, by = 2, length.out = 5)

# integer center, integer by, even length.out
# same as previous because even length.out is made odd by adding one
cseq(center = 10, by = 2, length.out = 4)

# double center, double by, odd length.out
# doubles also possible
cseq(center = 10.5, by = 2.25, length.out = 7)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S
Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[base]{seq}},
}
