% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{css_insert_wide_chunk}
\alias{css_insert_wide_chunk}
\title{Insert css code chunk to make consult report document wide}
\usage{
css_insert_wide_chunk(width_main = 15000, width_tocify = 400)
}
\arguments{
\item{width_main}{width in pixels of the main container}

\item{width_tocify}{width in pixes of the table of contents container}
}
\description{
Insert css code chunk to make consult report document wide
}
