% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_new_user}
\alias{db_new_user}
\title{Populate a tibble with new user information from LDAP}
\usage{
db_new_user(email, user_id, icts, app_role = c("user", "admin"))
}
\arguments{
\item{email}{mail of the user}

\item{user_id}{primary key of user}

\item{icts}{integer indicating icts membership status}

\item{app_role}{role of the user}
}
\value{
a tibble with populated user information
}
\description{
Populate a tibble with new user information from LDAP
}
