% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_next_user_consult_id}
\alias{db_next_user_consult_id}
\title{Next available user_consult_id}
\usage{
db_next_user_consult_id(cnf = "~/.my.cnf", group = "aws-rds-mysql80")
}
\arguments{
\item{cnf}{MySQL configuration file}

\item{group}{name of group in cnf file}
}
\value{
integer value of next available user_consult_id
}
\description{
Next available user_consult_id
}
