% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_search_consults}
\alias{db_search_consults}
\title{Run shiny app to search consult database}
\usage{
db_search_consults(
  cnf = "~/.my.cnf",
  group = "aws-rds-mysql80",
  consult_dir = Sys.getenv("WU_CONSULT_DIR")
)
}
\arguments{
\item{cnf}{MySQL configuration file}

\item{group}{name of group in cnf file}

\item{consult_dir}{consult directory}
}
\description{
Run shiny app to search consult database
}
