% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_ortho_consult}
\alias{db_sql_insert_ortho_consult}
\title{Construct SQL insert syntax for a new ortho consult}
\usage{
db_sql_insert_ortho_consult(
  consult_id,
  ortho_record_id,
  redcap_credentials = REDCapR::retrieve_credential_local(path_credential =
    "~/.REDCapR", project_id = 6546)
)
}
\arguments{
\item{consult_id}{id of the consult including version}

\item{ortho_record_id}{REDCap Ortho request identifier}

\item{redcap_credentials}{REDcap credentials}
}
\value{
SQL syntax to insert requisite data
}
\description{
Construct SQL insert syntax for a new ortho consult
}
\examples{
\dontrun{
db_sql_insert_ortho_consult("2020-11-04-1-v1", 7)
}
}
