% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user}
\alias{db_sql_insert_user}
\title{Construct SQL insert syntax for new users}
\usage{
db_sql_insert_user(email, user_id, icts, app_role)
}
\arguments{
\item{email}{mail of the user}

\item{user_id}{primary key of user}

\item{icts}{integer indicating icts membership status}

\item{app_role}{role of the user}
}
\value{
SQL syntax to insert new users
}
\description{
Construct SQL insert syntax for new users
}
