% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user_clip}
\alias{db_sql_insert_user_clip}
\title{Copy SQL insert syntax for new users to the clipboard}
\usage{
db_sql_insert_user_clip(email, user_id, icts, app_role)
}
\arguments{
\item{email}{mail of the user}

\item{user_id}{primary key of user}

\item{icts}{integer indicating icts membership status}

\item{app_role}{role of the user}
}
\description{
Copy SQL insert syntax for new users to the clipboard
}
\examples{
\dontrun{
db_sql_insert_user_clip(email = c("poo@wustl.edu", "piglet@wustl.edu"),
                        user_id = 1:2,
                        icts = c(1, 0),
                        app_role = c("admin", "user"))
}
}
