% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user_consult}
\alias{db_sql_insert_user_consult}
\title{Construct SQL insert syntax for new user-consult pairings}
\usage{
db_sql_insert_user_consult(user_consult_id, consult_id, user_id, role)
}
\arguments{
\item{user_consult_id}{primary key for table}

\item{consult_id}{consult id including version number}

\item{user_id}{foreign key for user from user table}

\item{role}{role of the user on the consult}
}
\value{
SQL syntax to insert new users
}
\description{
Construct SQL insert syntax for new user-consult pairings
}
