% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user_consult_clip}
\alias{db_sql_insert_user_consult_clip}
\title{Copy SQL insert syntax for new user-consult pairings to the clipboard}
\usage{
db_sql_insert_user_consult_clip(user_consult_id, consult_id, user_id, role)
}
\arguments{
\item{user_consult_id}{primary key for table}

\item{consult_id}{consult id including version number}

\item{user_id}{foreign key for user from user table}

\item{role}{role of the user on the consult}
}
\description{
Copy SQL insert syntax for new user-consult pairings to the clipboard
}
\examples{
\dontrun{
db_sql_insert_user_consult_clip(user_consult_id = 57:58, 
                                consult_id = "2020-11-01-1-v1", 
                                user_id = c(1, 14), 
                                role = c("statistician", "principal"))
}
}
