% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user_consult_single}
\alias{db_sql_insert_user_consult_single}
\title{Construct SQL insert syntax for a new user-consult pairing}
\usage{
db_sql_insert_user_consult_single(
  user_consult_id,
  consult_id,
  user_id,
  role = c("principal", "statistician", "student", "resident", "coinvestigator")
)
}
\arguments{
\item{user_consult_id}{primary key for table}

\item{consult_id}{consult id including version number}

\item{user_id}{foreign key for user from user table}

\item{role}{role of the user on the consult}
}
\value{
SQL syntax to insert a new user-consult pairing
}
\description{
Construct SQL insert syntax for a new user-consult pairing
}
