% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_sql_insert_user_single}
\alias{db_sql_insert_user_single}
\title{Construct SQL insert syntax for a new user}
\usage{
db_sql_insert_user_single(email, user_id, icts, app_role = "user")
}
\arguments{
\item{email}{mail of the user}

\item{user_id}{primary key of user}

\item{icts}{integer indicating icts membership status}

\item{app_role}{role of the user}
}
\value{
SQL syntax to insert a new user
}
\description{
Construct SQL insert syntax for a new user
}
