% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitools.R
\name{epitools_oddsratio}
\alias{epitools_oddsratio}
\title{Tidy odds ratio estimation and confidence intervals}
\usage{
epitools_oddsratio(data, outcome, exposure, na.rm = FALSE, ...)
}
\arguments{
\item{data}{A data frame or data frame extension (e.g. a tibble).}

\item{outcome}{Name of variable containing dichotomous outcome status.}

\item{exposure}{Name of exposure variable containing exposure status.}

\item{na.rm}{A logical value indicating whether to remove counted NA values.}

\item{...}{Arguments passed to \code{\link[epitools]{oddsratio}}.}
}
\value{
A tidy tibble of the output from \code{\link[epitools]{oddsratio}}.
}
\description{
Tidy odds ratio estimation and confidence intervals
}
\note{
If outcome and exposure are not of type factor, the function will
convert them to factors with default levels.
}
\examples{
## Case-control study assessing whether exposure to tap water
## is associated with cryptosporidiosis among AIDS patients
tapw <- c("Lowest", "Intermediate", "Highest")
outc <- c("Case", "Control")
dat <- matrix(c(2, 29, 35, 64, 12, 6), 3, 2, byrow = TRUE)
dimnames(dat) <- list("Tap water exposure" = tapw, "Outcome" = outc)
epitools::oddsratio(dat, rev = "c")

tribble(
  ~tapw, ~outc, ~n,
  'lowest', 'case', 2,
  'lowest', 'control', 29,
  'intermediate', 'case', 35,
  'intermediate', 'control', 64,
  'highest', 'case', 12,
  'highest', 'control', 6
) \%>\%
  tidyr::uncount() \%>\%
  mutate(
    tapw = factor(tapw, levels = c('lowest', 'intermediate', 'highest')),
    outc = factor(outc, levels = c('control', 'case'))
  ) \%>\%
  epitools_oddsratio(outc, tapw)
}
