% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitools.R
\name{epitools_riskratio}
\alias{epitools_riskratio}
\title{Tidy risk ratio estimation and confidence intervals}
\usage{
epitools_riskratio(.data, outcome, exposure, na.rm = FALSE, ...)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{outcome}{Name of variable containing dichotomous outcome status.}

\item{exposure}{Name of exposure variable containing exposure status.}

\item{na.rm}{A logical value indicating whether to remove counted NA values.}

\item{...}{Arguments passed to \code{\link[epitools]{riskratio}}.}
}
\value{
A tidy tibble of the output from \code{\link[epitools]{riskratio}}.
}
\description{
Tidy risk ratio estimation and confidence intervals
}
\note{
If outcome and exposure are not of type factor, the function will
convert them to factors with default levels.
}
\examples{
## Selvin 1998, p. 289
sel <- matrix(c(178, 79, 1411, 1486), 2, 2)
dimnames(sel) <- list(
  "Behavior type" = c("Type A", "Type B"),
  "Outcome" = c("CHD", "No CHD")
)
epitools::riskratio.boot(sel, rev = "b")

bind_rows(
  tibble(behavior = "Type B", chd = "no", n = 1486),
  tibble(behavior = "Type A", chd = "no", n = 1411),
  tibble(behavior = "Type B", chd = "yes", n = 79),
  tibble(behavior = "Type A", chd = "yes", n = 178)
) \%>\%
  tidyr::uncount() \%>\%
  mutate(
    behavior = factor(behavior, levels = c("Type B", "Type A")),
    chd = factor(chd, levels = c("no", "yes"))
  ) \%>\%
  epitools_riskratio(chd, behavior)
}
\references{
Steve Selvin (1998), Modern Applied Biostatistical Methods Using S-Plus, 1st Edition, Oxford University Press
}
