% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{es_cohens_d}
\alias{es_cohens_d}
\title{Cohen's d}
\usage{
es_cohens_d(m_0, m_1, s_0, s_1)
}
\arguments{
\item{m_0}{mean of group 0}

\item{m_1}{mean of group 1}

\item{s_0}{standard deviation of group 0}

\item{s_1}{standard deviation of group 1}
}
\value{
Cohen's d
}
\description{
Cohen's d
}
\examples{
mtcars \%>\%
  group_by(vs) \%>\%
  summarize(m = mean(mpg), s = sd(mpg)) \%>\%
  tidyr::pivot_wider(names_from = vs, values_from = c(m, s)) \%>\%
  rowwise() \%>\%
  mutate(cohens_d = es_cohens_d(m_0, m_1, s_0, s_1))

}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences.
Routledge.
}
