% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{es_glass_delta}
\alias{es_glass_delta}
\title{Glass' Delta}
\usage{
es_glass_delta(m_0, m_1, s)
}
\arguments{
\item{m_0}{mean of group 0}

\item{m_1}{mean of group 1}

\item{s}{standard deviation of control group}
}
\value{
Glass' delta
}
\description{
Glass' Delta
}
\examples{
mtcars \%>\%
  group_by(vs) \%>\%
  summarize(m = mean(mpg), s = sd(mpg)) \%>\%
  tidyr::pivot_wider(names_from = vs, values_from = c(m, s)) \%>\%
  rowwise() \%>\%
  mutate(glass_delta = es_glass_delta(m_0, m_1, s_0))

}
\references{
Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Academic Press.
}
