% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{es_hedges_g}
\alias{es_hedges_g}
\title{Hedges g}
\usage{
es_hedges_g(n_0, n_1, m_0, m_1, s_0, s_1)
}
\arguments{
\item{n_0}{number of observations in group 0}

\item{n_1}{number of observations in group 1}

\item{m_0}{mean of group 0}

\item{m_1}{mean of group 1}

\item{s_0}{standard deviation of group 0}

\item{s_1}{standard deviation of group 1}
}
\value{

}
\description{
Hedges g
}
\examples{
mtcars \%>\% 
  group_by(vs) \%>\% 
  summarize(n = n(), m = mean(mpg), s = sd(mpg)) \%>\% 
  tidyr::pivot_wider(names_from = vs, values_from = c(n, m, s)) \%>\% 
  rowwise() \%>\% 
  mutate(hedges_g = es_hedges_g(n_0, n_1, m_0, m_1, s_0, s_1))

}
\references{
Hedges, L. V. (1981). Distribution theory for Glass' estimator of effect size and related estimators. Journal of Educational Statistics, 6(2), 107-128. https://doi.org/10.3102/10769986006002107
}
