% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{es_map}
\alias{es_map}
\title{Map between various effect sizes}
\usage{
es_map(
  size,
  type = c("r", "R^2", "d", "or", "f", "f^2", "eta^2", "omega^2", "cl")
)
}
\arguments{
\item{size}{effect size to convert (see details)}

\item{type}{the effect type to convert}
}
\value{
list containing converted effects
}
\description{
Map between various effect sizes
}
\details{
Ported from Microsoft Excel Spreadsheet created by Jamie DeCoster
on 2012-06-19 \url{http://www.stat-help.com} with the addition of R^2, f^2,
and omega^2.

Expected values for type include: \cr
\tabular{ll}{
r       \tab \href{https://en.wikipedia.org/wiki/Effect_size#Pearson_r_or_correlation_coefficient}{Pearson's r}          \cr
R^2     \tab \href{https://en.wikipedia.org/wiki/Effect_size#Coefficient_of_determination}{Coefficient of Determination} \cr
d       \tab \href{https://en.wikipedia.org/wiki/Effect_size#Cohen's_d}{Cohen's d}                                       \cr
or      \tab \href{https://en.wikipedia.org/wiki/Effect_size#Odds_ratio}{Odds ratio}                                     \cr
f       \tab \href{https://en.wikipedia.org/wiki/Effect_size#Cohen's_f2}{Cohen's f}                                      \cr
f^2     \tab \href{https://en.wikipedia.org/wiki/Effect_size#Cohen's_f2}{Cohen's f^2}                                    \cr
eta^2   \tab \href{https://en.wikipedia.org/wiki/Effect_size#Eta-squared_(e2)}{Eta-squared}                              \cr
omega^2 \tab \href{https://en.wikipedia.org/wiki/Effect_size#Omega-squared_(w2)}{Omega-squared}                          \cr
cl      \tab \href{https://en.wikipedia.org/wiki/Effect_size#Common_language_effect_size}{Common Language Effect Size}
}
}
\examples{
# small Cohen's d
es_map(0.20, "d")

# medium Cohen's d
es_map(0.50, "d")

# large Cohen's d
es_map(0.80, "d")

# testing code
sapply(
  unlist(
    as.list(
      formals(es_map)$type[2:10]
    )
  ),
  function(x) es_map(0.10, x)
)
}
\references{
Formulas for converting between f, f-squared, r-squared, eta-squared, and d
were taken from: \cr
Cohen J. (1988). Statistical Power Analysis for the Behavioral Sciences
(2nd ed.), Hillsdale, NJ: Erlbaum. pp. 281, 284, 285

Formulas for converting between r and d were taken from: \cr
Rosenthal, R. (1994). Parametric measures of effect size. In H. Cooper & L.
V. Hedges (Eds.), The Handbook of Research Synthesis. New York, NY: Sage.
pp. 239.

Formulas for converting between the odds ratio and d were taken from: \cr
Borenstein, M., Hedges, L. V., Higgins, J. P. T., & Rothstein, H. R. (2009).
Introduction to Meta-Analysis. Chichester, West Sussex, UK: Wiley.

Formulas for converting between the cl (also called the "Common Language
Effect Size") and d were taken from: \cr
Ruscio, J. (2008). A probability-based measure of effect size: Robustness
to base rates and other factors. Psychological Methods, 13, 19-30.

Formulas for converting between f-squared and omega-squared were taken
from: \cr
Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals
and tests of close fit in the analysis of variance and contrast analysis.
Psychological Methods, 9(2), 164-182.
}
