% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{estimate_document}
\alias{estimate_document}
\title{Estimate document
Format for creating a WashU estimate on School of Medicine letterhead}
\usage{
estimate_document(
  template = find_resource("template_resource", "estimate", "template.tex"),
  latex_engine = "xelatex",
  includes,
  ...
)
}
\arguments{
\item{template}{Pandoc template to use for rendering. Passed value ignored
in favor of default.}

\item{latex_engine}{LaTeX engine for producing PDF output. Passed value
ignored in favor of default.}

\item{includes}{Named list of additional content to include within the
document (typically created using the \code{\link[rmarkdown]{includes}} function).}

\item{...}{Arguments to \code{\link[rmarkdown]{pdf_document}}}
}
\description{
Estimate document
Format for creating a WashU estimate on School of Medicine letterhead
}
