% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{extract_directory_field}
\alias{extract_directory_field}
\title{Extract directory field}
\usage{
extract_directory_field(v, k)
}
\arguments{
\item{v}{character vector intermixing keys and values for a single result
from a directory search}

\item{k}{key to search for}
}
\value{
desired value or NA
}
\description{
Extract a field from a WashU directory search
}
