% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{find_package_font_path}
\alias{find_package_font_path}
\title{Find package font path}
\usage{
find_package_font_path()
}
\value{
Absolute path to the package font directory
}
\description{
Find package font path
}
