% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{find_resource}
\alias{find_resource}
\title{Find resource}
\usage{
find_resource(
  type = c("template_resource", "template_skeleton", "global_resource",
    "global_font_path"),
  fork,
  file
)
}
\arguments{
\item{type}{type of resource to find}

\item{fork}{specify which subdirectory to search in the overall type path}

\item{file}{file to locate}
}
\value{
absolute path to the resource
}
\description{
Find resource
}
