% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{find_template_resource}
\alias{find_template_resource}
\title{Find template resource}
\usage{
find_template_resource(template, file = "template.tex")
}
\arguments{
\item{template}{template to look under}

\item{file}{file to search for in the given template resources directory}
}
\value{
absolute path to the desired file
}
\description{
Find template resource
}
