% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{find_template_skeleton}
\alias{find_template_skeleton}
\title{Find template skeleton}
\usage{
find_template_skeleton(template)
}
\arguments{
\item{template}{template to look under}
}
\value{
absolute path to the template skeleton
}
\description{
Find template skeleton
}
