% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%==\%}
\alias{\%==\%}
\title{Custom compare}
\usage{
lhs \%==\% rhs
}
\arguments{
\item{lhs}{left-hand side}

\item{rhs}{right-hand side}
}
\value{
logical indicating match
}
\description{
Custom compare
}
\details{
Compare as with ==, but NA == NA returns TRUE
}
