% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%missing\%}
\alias{\%missing\%}
\title{Backup value for missing arguments}
\usage{
x \%missing\% y
}
\arguments{
\item{x}{argument to check for missing}

\item{y}{value to use in case of missing}
}
\description{
Backup value for missing arguments
}
