% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In
Return a logical vector indicating if there is not a match or is for its left operand}
\usage{
x \%notin\% table
}
\arguments{
\item{x}{vector or \code{NULL}: the values to be matched.
    \link[base]{Long vectors} are supported.}

\item{table}{vector or \code{NULL}: the values to be matched against.
    \link[base]{Long vectors} are not supported.}
}
\value{
A logical vector, indicating if a match was not located for each element of x: thus the values are TRUE or FALSE and never NA
}
\description{
Not In
Return a logical vector indicating if there is not a match or is for its left operand
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
\seealso{
\code{\link[base]{match}}
}
