% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hr.R
\name{hr.test}
\alias{hr.test}
\title{Hazard ratio test}
\usage{
hr.test(beta1, se1, beta2, se2, r = 0, n = 10000)
}
\arguments{
\item{beta1}{beta estimate for the first hazard ratio}

\item{se1}{standard error of beta estimate for the first hazard ratio}

\item{beta2}{standard error of beta estimate for the second hazard ratio}

\item{se2}{standard error of beta estimate for the second hazard ratio}

\item{r}{correlation between the first and second risk scores (i.e., predictors)}

\item{n}{number of samples from which the hazard ratios were estimated}
}
\value{
A list with class "htest" containing the following components:
\tabular{ll}{
statistic \tab the value of the t-statistic. \cr
parameter \tab the degrees of freedom for the t-statistic. \cr
p.value \tab the p-value for the test. \cr
conf.int \tab a confidence interval for the mean appropriate to the specified alternative hypothesis. \cr
estimate \tab the estimated mean or difference in means depending on whether it was a one-sample test or a two-sample test. \cr
null.value \tab the specified hypothesized value of the mean or mean difference depending on whether it was a one-sample test or a two-sample test. \cr
stderr \tab the standard error of the mean (difference), used as denominator in the t-statistic formula. \cr
alternative \tab a character string describing the alternative hypothesis. \cr
method \tab a character string indicating what type of t-test was performed. \cr
data.name \tab a character string giving the name(s) of the data.
}
}
\description{
Hazard ratio test
}
\note{
Currently only performs a directional test
}
\examples{
set.seed(12345)
age <- as.numeric(rnorm(100, 50, 10) >= 50)
size <- as.numeric(rexp(100, 1) > 1)
stime <- rexp(100)
cens <- runif(100, .5, 2)
sevent <- as.numeric(stime <= cens)
stime <- pmin(stime, cens)
coxm1 <- survival::coxph(survival::Surv(stime, sevent) ~ age)
coxm2 <- survival::coxph(survival::Surv(stime, sevent) ~ size)
hr.test(
  coxm1$coefficients,
  drop(sqrt(coxm1$var)),
  coxm2$coefficients,
  drop(sqrt(coxm2$var)),
  cor(age, size),
  100
)
}
\seealso{
\code{\link[survcomp]{hr.comp2}}
}
