% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{icts_directory}
\alias{icts_directory}
\title{ICTS membership directory search}
\usage{
icts_directory(query)
}
\arguments{
\item{query}{search query}
}
\value{
tibble of results
}
\description{
ICTS membership directory search
}
