% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_bootstrap}
\alias{impute_bootstrap}
\title{Replace missing values}
\usage{
impute_bootstrap(.data, x)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{x}{name of variable to impute on}
}
\value{
an object of the same type as .data with missing values in the desired column imputed by random sample from observed values
}
\description{
Replace missing values
}
\examples{
impute_bootstrap(airquality, Ozone)
}
\seealso{
\link[tidyimpute]{impute}
}
