% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{knit_print.washu_estimate_items}
\alias{knit_print.washu_estimate_items}
\title{Custom S3 Print for washu_estimate_items}
\usage{
\method{knit_print}{washu_estimate_items}(x, ...)
}
\arguments{
\item{x}{A washu_estimate_items object to print}

\item{...}{additional arguments}
}
\description{
Custom S3 Print for washu_estimate_items
}
