% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{knit_print.washu_table1}
\alias{knit_print.washu_table1}
\title{Custom S3 Print for washu_table1 objects}
\usage{
\method{knit_print}{washu_table1}(x, ...)
}
\arguments{
\item{x}{A washu_table1 object to print}

\item{...}{additional arguments}
}
\description{
Custom S3 Print for washu_table1 objects
}
