% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ldap_as_tibble}
\alias{ldap_as_tibble}
\title{Coerce LDAP content to data frame}
\usage{
ldap_as_tibble(ldap_content)
}
\arguments{
\item{ldap_content}{content to be coerced}
}
\value{
tibble representing the content
}
\description{
Coerces a character vector of LDAP content containing key-value, colon-delimited pairs into a wide format tibble
}
\seealso{
\code{\link{ldap_content}}
}
