% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ldap_clean_tibble}
\alias{ldap_clean_tibble}
\title{LDAP clean tibble}
\usage{
ldap_clean_tibble(ldap_tibble)
}
\arguments{
\item{ldap_tibble}{tibble to clean}
}
\value{
cleaned tibble
}
\description{
Normalize LDAP tibble data by re-storing data in tab-delimited variables
}
\details{
Some LDAP queries tested so far (i.e., sAMAccountName, mail, userPrincipalName) have objectClass
data stored in the DN variable, and memberOf data is stored as "\\n\\tmemberOf: "-delimited variable strings.
This function extracts and stores this data in tab-delimited variable strings.
}
\seealso{
\code{\link{ldap_as_tibble}}
}
