% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{ldap_content}
\alias{ldap_content}
\title{LDAP content}
\usage{
ldap_content(response, delim = "\\n\\n\\t")
}
\arguments{
\item{response}{curl response object}

\item{delim}{delimiter}
}
\value{
character vector representation of content
}
\description{
Coerces ldap curl response content to character vector by splitting on "\\n\\n\\t"
}
\seealso{
\code{\link{rm_raw_nul}}, \code{\link[base:rawConversion]{rawToChar}}, \code{\link[curl]{curl_fetch}}
}
