% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{letter_of_support_body}
\alias{letter_of_support_body}
\title{Letter of support body}
\usage{
letter_of_support_body(
  title,
  template = Sys.getenv("WU_LETTER_OF_SUPPORT_BODY"),
  pattern = "\\\\$title\\\\$"
)
}
\arguments{
\item{title}{title of the project to be supported}

\item{template}{path to letter of support body template provided by user}

\item{pattern}{pattern in the file to replace with title}
}
\description{
Letter of support body
}
