% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{letter_runtime_header_additions}
\alias{letter_runtime_header_additions}
\title{Letter runtime header additions}
\usage{
letter_runtime_header_additions()
}
\value{
path to a temporary header file
}
\description{
Letter runtime header additions
}
\details{
Package installation paths differ from system to system making
the location of some package resources unknown. Examples include the
location of package font files and the WashU SOM logo. This function
generates and writes the required tex to a temporary file and returns
the path to that file to be used as a header inclusion.
}
