% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{map}
\alias{map}
\title{Purrr-like map}
\usage{
map(.x, .f, ...)
}
\arguments{
\item{.x}{A list or atomic vector}

\item{.f}{A function to apply}

\item{...}{additional arguments}
}
\description{
Purrr-like map
}
\note{
implemented to avoid dependence on purrr
}
