% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{match_arg}
\alias{match_arg}
\title{Default in the event of NULL arguments}
\usage{
match_arg(x, default)
}
\arguments{
\item{x}{argument to check if null}

\item{default}{value to supply if argument is null}
}
\value{
original argument if not null; default otherwise
}
\description{
Default in the event of NULL arguments
}
