% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_auto_coalesce}
\alias{mutate_auto_coalesce}
\title{Auto coalesce after a join}
\usage{
mutate_auto_coalesce(data, suffix = c(".x", ".y"))
}
\arguments{
\item{data}{the data resulting from a join operation}

\item{suffix}{the suffixes used to disambiguate non-joined duplicate variables during the join}
}
\value{
a dataframe with duplicate variables coalesced and the original variants dropped
}
\description{
When joining data, there are often column conflicts. Coalescing is the most
common solution. This function attempts to identify the conflicting column
names after a join, coalesces them into the base name, and drops the
suffixed versions.
}
\examples{
x <- data.frame(id = 1:3, v = paste0('x', 1:3))
y <- data.frame(id = 2:4, v = paste0('y', 2:4))
d <- dplyr::left_join(x, y, by = 'id')
mutate_auto_coalesce(d)
}
