% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate.R
\name{mutate_dicotomize}
\alias{mutate_dicotomize}
\title{Create a dicotomized x column based on first rpart split}
\usage{
mutate_dicotomize(.data, x, y)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr)}

\item{x}{name of predictor to dicotomize}

\item{y}{name of response to base dicotomization on}
}
\value{
.data with a new factor representing the optimal split
}
\description{
Create a dicotomized x column based on first rpart split
}
\examples{
mutate_dicotomize(iris, Sepal.Length, Sepal.Width)

}
\seealso{
\link[rpart]{rpart}
}
