% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{parse_directory_entry}
\alias{parse_directory_entry}
\title{Parse Directory Entry}
\usage{
parse_directory_entry(e)
}
\arguments{
\item{e}{html node list representing a single result from a directory search}
}
\value{
data frame representing a single result from a directory search
}
\description{
Parse Directory Entry
}
