% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{pp_d}
\alias{pp_d}
\title{Power plot for Cohen's d and dz}
\usage{
pp_d(
  data,
  n,
  type = c("two.sample", "one.sample", "paired"),
  alternative = c("two.sided", "less", "greater"),
  sig.level = 0.05,
  min_pwr = NULL,
  effect = .data$d,
  labels = .data$d
)
}
\arguments{
\item{data}{tibble of effect size data}

\item{n}{sample sizes per sample for which to compute power}

\item{type}{type of t test : one- two- or paired-samples}

\item{alternative}{a character string specifying the alternative hypothesis, must be one of "two.sided" (default), "greater" or "less"}

\item{sig.level}{significance level (Type I error probability)}

\item{min_pwr}{minimum desirable power to label (NULL for no label)}

\item{effect}{variable in data representing the effect sizes}

\item{labels}{effect labels to be used for legend}
}
\value{
object of class "ggplot"
}
\description{
Power plot for Cohen's d and dz
}
\note{
The effect variable is converted to factor in the given order and labels applied respectively.
}
\examples{
## Exercise 2.1 P. 40 from Cohen (1988)
dplyr::tibble(d = .50,
              y = "learning",
              x = "opportunity",
              source = "Cohen (1988)") \%>\% 
  pp_d(15:45,
       labels = sprintf("\%s \%s", format(round(d, 2), nsmall = 2), source))
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[pwr]{pwr.t.test}}
}
