% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power.R
\name{pp_f2}
\alias{pp_f2}
\title{Power plot for Cohen's f2}
\usage{
pp_f2(
  data,
  N,
  u1 = 1,
  u2 = 0,
  sig.level = 0.05,
  min_pwr = NULL,
  effect = .data$f2,
  labels = .data$f2
)
}
\arguments{
\item{data}{tibble of effect size data}

\item{N}{total sample sizes for which to compute power}

\item{u1}{degrees of freedom for effect generator}

\item{u2}{degrees of freedom for other predictors (1 + u1 + u2 + v = N)}

\item{sig.level}{significance level (Type I error probability)}

\item{min_pwr}{minimum desirable power to label (NULL for no label)}

\item{effect}{variable in data representing the effect sizes}

\item{labels}{effect labels to be used for legend}
}
\value{
object of class "ggplot"
}
\description{
Power plot for Cohen's f2
}
\note{
The effect variable is converted to factor in the given order and labels applied respectively.
}
\examples{
## Exercise 9.1 P. 424 from Cohen (1988)
dplyr::tibble(
  f2 = 0.1 / (1 - 0.1),
  y = "sales",
  x = "demographics",
  source = "Cohen (1988)"
) \%>\%
  pp_f2(85:105,
    u1 = 5,
    labels = sprintf("\%s \%s", format(round(f2, 2), nsmall = 2), source)
  )
}
\references{
Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd ed.). Hillsdale,NJ: Lawrence Erlbaum.
}
\seealso{
\code{\link[ggplot2]{ggplot}}, \code{\link[pwr]{pwr.f2.test}}
}
