% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{print.dtest}
\alias{print.dtest}
\title{Print Objects of Type dtest}
\usage{
\method{print}{dtest}(
  x,
  format = c("tibble", "text"),
  digits = 4,
  digits.df = 0,
  digits.stat = digits,
  digits.p = digits,
  ...
)
}
\arguments{
\item{x}{an object of class "dtype", a result of a call to
\code{\link{deviance_test}}}

\item{format}{format of the output}

\item{digits}{if format is text, default decimal digits to print}

\item{digits.df}{if format is text, df decimal digits to print}

\item{digits.stat}{if format is text, test statistic decimal digits to print}

\item{digits.p}{if format is text, p-value decimal digits to print}

\item{...}{additional parameters passed to other functions}
}
\description{
Print Objects of Type dtest
}
