% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{print.glm}
\alias{print.glm}
\title{Print Objects of Type glm}
\usage{
\method{print}{glm}(x, format = c("default", "text"), sig.level = 0.05, ...)
}
\arguments{
\item{x}{an object of class "glm", a result of a call to
\code{\link[stats]{glm}}}

\item{format}{format of the output}

\item{sig.level}{significance level}

\item{...}{additional parameters passed to other functions}
}
\description{
Print Objects of Type glm
}
