% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{print.HLtest}
\alias{print.HLtest}
\title{Print Objects of Type HLtest}
\usage{
\method{print}{HLtest}(
  x,
  format = c("default", "tibble", "text"),
  digits = 4,
  digits.df = 0,
  digits.stat = digits,
  digits.p = digits,
  ...
)
}
\arguments{
\item{x}{an object of class "HLtest", a result of a call to \code{\link[HosmerLemeshow]{vcdExtra::HosmerLemeshow()}}}

\item{format}{format of the output}

\item{digits}{if format is text, default decimal digits to print}

\item{digits.df}{if format is text, df decimal digits to print}

\item{digits.stat}{if format is text, test statistic decimal digits to print}

\item{digits.p}{if format is text, p-value decimal digits to print}

\item{...}{additional parameters passed to other functions}
}
\description{
Print Objects of Type HLtest
}
