% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{print.r_sqr}
\alias{print.r_sqr}
\title{Print Objects of Type r_sqr}
\usage{
\method{print}{r_sqr}(x, format = c("default", "tibble", "text"), digits.r2 = 4, ...)
}
\arguments{
\item{x}{an object of class "r_sqr", a result of a call to \code{\link[r2]{washu::r2()}}}

\item{format}{format of the output}

\item{digits.r2}{if format is text, default decimal digits to print}

\item{...}{additional parameters passed to other functions}
}
\description{
Print Objects of Type r_sqr
}
