% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{print.washu_table1}
\alias{print.washu_table1}
\title{Print washu_table1 objects}
\usage{
\method{print}{washu_table1}(x)
}
\arguments{
\item{x}{an object of type washu_table1}
}
\value{
the argument x, invisibly
}
\description{
Print washu_table1 objects
}
