% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{print_rgb}
\alias{print_rgb}
\title{Print RGB color}
\usage{
print_rgb(col)
}
\arguments{
\item{col}{vector of any of the three kinds of R color specifications (i.e.,
name, hex, or integer)}
}
\value{
character string representation of the rgb color
}
\description{
Print a color as an RGB string
}
\examples{
\dontrun{
print_rgb(wu_colors$red)
}
}
