% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{protect_tex_input}
\alias{protect_tex_input}
\title{Protect TeX input}
\usage{
protect_tex_input(x, ...)
}
\arguments{
\item{x}{TeX to escape}

\item{...}{additional arguments}
}
\description{
Protect TeX input
}
