% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{push_slashes}
\alias{push_slashes}
\title{Push slashes}
\usage{
push_slashes(x)
}
\arguments{
\item{x}{text to modify}
}
\value{
x with backslashes replaced with forward slashes
}
\description{
Push slashes
}
