% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{r2}
\alias{r2}
\title{Compute the model's R2}
\usage{
r2(model, ...)
}
\arguments{
\item{model}{A statistical model.}

\item{...}{Arguments passed down to the related r2-methods.}
}
\description{
Compute the model's R2
}
\note{
This is a wrapper around \code{\link[r2]{performance::r2()}} for a consistent class for printing
}
