% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{recover_data_viewer_cache_objects}
\alias{recover_data_viewer_cache_objects}
\title{Recover Data Viewer Cache Objects}
\usage{
recover_data_viewer_cache_objects()
}
\value{
a list containing recovered utils::View() objects
}
\description{
Recover Data Viewer Cache Objects
}
\details{
If the current R session is terminated while an object is being
viewed with utils::View(), RStudio will cache the object, meaning the object
can be recovered.
}
\references{
https://support.rstudio.com/hc/en-us/articles/200534577-Resetting-RStudio-Desktop-s-State
}
