% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{redcap_export_project_information}
\alias{redcap_export_project_information}
\title{Export Project Information}
\usage{
redcap_export_project_information(
  token,
  redcap_uri = "https://redcap.wustl.edu/redcap/api/",
  format = c("xml", "csv", "json"),
  returnFormat = c("xml", "csv", "json")
)
}
\arguments{
\item{token}{The API token specific to your REDCap project and username (each
token is unique to each user for each project). See the section on the
left-hand menu for obtaining a token for a given project.}

\item{redcap_uri}{The URI (uniform resource identifier) of the REDCap
project.}

\item{format}{xml \[default\], csv, json}

\item{returnFormat}{xml \[default\], csv, json}
}
\value{
httr::response() object containing the project information
}
\description{
Export Project Information
}
