% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{redcap_logical}
\alias{redcap_logical}
\title{REDCap Logical}
\usage{
redcap_logical(name, value)
}
\arguments{
\item{name}{name to give logical}

\item{value}{R logical to cast as REDCap API logical}
}
\value{
an R logical represented as a named list of lowercase character
}
\description{
REDCap Logical
}
\examples{
\dontrun{
a <- TRUE
redcap_logical("a", a)

b <- FALSE
redcap_logical("b", b)

c <- "not gonna work"
redcap_logical("c", c)
}
}
