% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcap.R
\name{redcap_migrate_project}
\alias{redcap_migrate_project}
\title{Migrate a Project}
\usage{
redcap_migrate_project(
  redcap_uri_src = "https://redcap.wustl.edu/redcap/srvrs/prod_v3_1_0_001/redcap/api/",
  redcap_uri_dst = "https://redcap.wustl.edu/redcap/api/",
  token_src,
  token_spr
)
}
\arguments{
\item{redcap_uri_src}{The URI (uniform resource identifier) of the source
REDCap project.}

\item{redcap_uri_dst}{The URI (uniform resource identifier) of the
destination REDCap project.}

\item{token_src}{The API token specific to your source REDCap project and
username (each token is unique to each user for each project). See the
section on the left-hand menu for obtaining a token for a given project.}

\item{token_spr}{The Super API Token specific to a user}
}
\description{
Migrate a Project
}
\examples{
\dontrun{
washu::redcap_migrate_project(
  redcap_uri_src =
  "https://redcap.wustl.edu/redcap/srvrs/prod_v3_1_0_001/redcap/api/",
  redcap_uri_dst =
  "https://redcap.wustl.edu/redcap/api/",
  token_src = my_source_token,
  token_spr = my_super_user_token
)
}
}
