% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redcapr.R
\name{retrieve_credential_local}
\alias{retrieve_credential_local}
\title{Read a token and other credentials from a (non-REDCap)
database or file}
\usage{
retrieve_credential_local(project_id, path_credential = "~/.REDCapR", ...)
}
\arguments{
\item{project_id}{The ID assigned to the project withing REDCap.  This
allows the user to store tokens to multiple REDCap projects in one file.
Required}

\item{path_credential}{The file path to the CSV containing the credentials.
Required.}
}
\value{
A list of the following elements:
\itemize{
\item \code{redcap_uri}: The URI of the REDCap Server.
\item \code{username}: Username.
\item \code{project_id}: The ID assigned to the project withing REDCap.
\item \code{token}: The token to pass to the REDCap server
\item \code{comment}: An optional string.
}
}
\description{
These functions are not essential to calling the REDCap API,
but instead are functions that help manage tokens securely.
}
\details{
If the database elements are created with the script provided in package's
'Security Database' vignette, the default values will work.
}
\examples{
# ---- Local File Example ----------------------------
path <- system.file("misc/example.credentials", package = "REDCapR")
(p1  <- REDCapR::retrieve_credential_local(path, 153L))
(p2  <- REDCapR::retrieve_credential_local(path, 212L))
}
\author{
Will Beasley
}
