% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{rgb_value_pad}
\alias{rgb_value_pad}
\title{RGB value pad}
\usage{
rgb_value_pad(x)
}
\arguments{
\item{x}{integer vector of length 3 with integers between 0 and 255 inclusive}
}
\value{
character vector with three characters per element
}
\description{
Pad RGB values with leading 0's
}
\examples{
\dontrun{
rgb_value_pad(c(0, 10, 100))
}
}
