% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{rm_raw_nul}
\alias{rm_raw_nul}
\title{Remove raw nul}
\usage{
rm_raw_nul(data)
}
\arguments{
\item{data}{content element from a curl response}
}
\value{
raw content with the nul values deleted
}
\description{
Removes nul values from curl response content to allow conversion to a
character representation
}
\seealso{
\code{\link{ldap_content}},
\code{\link[base:rawConversion]{rawToChar}}, \code{\link[curl]{curl_fetch}}
}
