% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_connect}
\alias{saturn_connect}
\title{Saturn Connect}
\usage{
saturn_connect(
  user = Sys.getenv("WUSTL_KEY_USER"),
  passwd = Sys.getenv("WUSTL_KEY_PASS"),
  verbose = FALSE
)
}
\arguments{
\item{user}{saturn username as string}

\item{passwd}{either a string or a callback function for password prompt}

\item{verbose}{either TRUE/FALSE or a value between 0 and 4 indicating log level:
0: no logging, 1: only warnings, 2: protocol, 3: packets or 4: full stack trace.}
}
\value{
ssh session
}
\description{
Saturn Connect
}
\examples{
\dontrun{
# retrive user and pass from .Renviron file in user home directory
session <- saturn_connect()
}

}
\seealso{
\code{\link[ssh]{ssh_connect}}
}
