% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_download}
\alias{saturn_download}
\title{Saturn Download}
\usage{
saturn_download(session, files, to = ".", verbose = FALSE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{files}{path to files or directory to transfer}

\item{to}{existing directory on the destination where \code{files} will be copied into}

\item{verbose}{print progress while copying files}
}
\description{
Saturn Download
}
\examples{
\dontrun{
# recursively download files and directories
session <- saturn_connect()
saturn_download(session, "~/target/*", tempdir())
}

}
\seealso{
\code{\link[ssh]{scp_download}}
}
