% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_execute}
\alias{saturn_execute}
\title{Saturn Execute}
\usage{
saturn_execute(session, command = "whoami", error = TRUE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{command}{The command or script to execute}

\item{error}{automatically raise an error if the exit status is non-zero}
}
\value{
list containing exit status, buffered raw stdout, and buffered raw stderr
}
\description{
Saturn Execute
}
\examples{
\dontrun{
session <- saturn_connect()
out <- saturn_execute(session)
rawToChar(out$stdout)
}

}
\seealso{
\code{\link[ssh]{ssh_exec_internal}}
}
