% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_download}
\alias{saturn_job_download}
\title{Saturn Job Download}
\usage{
saturn_job_download(session, job_name, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{job_name}{job name}

\item{...}{additional arguments passed to \code{\link{saturn_upload}}}
}
\description{
Saturn Job Download
}
\examples{
\dontrun{
session <- saturn_connect()
job_name <- "my_job"
saturn_job_init(job_name, "pooh")
saturn_job_upload(session, job_name)
saturn_job_submit(session, job_name)
saturn_job_running(session, job_name)
saturn_job_download(session, job_name)
}
}
