% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_edit_source}
\alias{saturn_job_edit_source}
\title{Saturn Job Edit Source}
\usage{
saturn_job_edit_source(job_name)
}
\arguments{
\item{job_name}{job name}
}
\description{
Saturn Job Edit Source
}
\examples{
\dontrun{
job_name <- "my_job"
saturn_job_init(job_name, edit_script = FALSE)
saturn_job_edit_source(job_name)
}
}
