% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_remove}
\alias{saturn_job_remove}
\title{Saturn Job Remove}
\usage{
saturn_job_remove(session, job_name, location = c("remote", "local"), ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{job_name}{job name}

\item{location}{remove the job directory on the cluster or locally}

\item{...}{additional arguments passed to \code{\link{saturn_execute}}, \code{\link[base]{unlink}}}
}
\description{
Saturn Job Remove
}
\examples{
\dontrun{
session <- saturn_connect()
job_name <- "my_job"
saturn_job_init(job_name)
saturn_job_upload(session, job_name)
saturn_job_submit(session, job_name)
saturn_job_running(session, job_name)
saturn_job_download(session, job_name)
saturn_job_remove(session, job_name)
saturn_job_remove(session, job_name, "local")
}
}
