% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_running}
\alias{saturn_job_running}
\title{Saturn Job Running}
\usage{
saturn_job_running(session, job_name)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{job_name}{job name}
}
\value{
logical indicating if job is running
}
\description{
Saturn Job Running
}
\examples{
\dontrun{
session <- saturn_connect()
job_name <- "my_job"
saturn_job_init(job_name, "pooh")
saturn_job_upload(session, job_name)
saturn_job_submit(session, job_name)
saturn_job_running(session, job_name)
}
}
