% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_schedule}
\alias{saturn_job_schedule}
\title{Saturn Job Schedule}
\usage{
saturn_job_schedule(session, job_name, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{job_name}{job name}

\item{...}{additional arguments passed to \code{\link{saturn_execute}}, \code{\link{saturn_upload}}}
}
\description{
Saturn Job Schedule
}
\details{
\tabular{ll}{
\strong{Command}         \tab \strong{Details}               \cr
crontab -e               \tab create and edit a crontab file \cr
crontab -l               \tab view contents of crontab file  \cr
crontab -r               \tab unschedule all jobs            \cr
crontab -a job_name.cron \tab schedule a job
}
}
\note{
check mail with mail command
}
\examples{
\dontrun{
session <- saturn_connect()
job_name <- "my_job"
saturn_job_init(job_name, add_cron = TRUE)
saturn_job_upload(session, job_name)
saturn_job_schedule(session, job_name)
}

}
