% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_upload}
\alias{saturn_job_upload}
\title{Saturn Job Upload}
\usage{
saturn_job_upload(session, job_name, mk_spath = FALSE, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{job_name}{job name}

\item{mk_spath}{make the spath if it does not exist}

\item{...}{additional arguments passed to \code{\link{saturn_execute}} and \code{\link{saturn_upload}}}
}
\description{
Saturn Job Upload
}
\examples{
\dontrun{
session <- saturn_connect()
job_name <- "my_job"
saturn_job_init(job_name, "pooh")
saturn_job_upload(session, job_name)
}
}
