% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_pbsnodes}
\alias{saturn_pbsnodes}
\title{Saturn pbsnodes}
\usage{
saturn_pbsnodes(session, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{...}{additional arguments passed to \code{\link{saturn_qstat}}}
}
\value{
status of all compute nodes
}
\description{
Saturn pbsnodes
}
\examples{
\dontrun{
session <- saturn_connect()
r <- saturn_pbsnodes(session)
cat(rawToChar(r$stdout))
}
}
