% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_qstat}
\alias{saturn_qstat}
\title{Saturn qstat Command}
\usage{
saturn_qstat(session, args, ...)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{args}{arguments to pass to qstat}

\item{...}{additional arguments passed to \code{\link{saturn_execute}}}
}
\description{
Saturn qstat Command
}
\details{
\tabular{ll}{
\strong{Argument} \tab \strong{Details}           \cr
-q        \tab list all queues                    \cr
-a        \tab list all jobs                      \cr
-u userid \tab list jobs for userid               \cr
-r        \tab list running jobs                  \cr
-f job_id \tab list full information about job_id \cr
-Qf queue \tab list full information about queue  \cr
-B        \tab list summary status of the job server
}
}
\examples{
\dontrun{
session <- saturn_connect()
r <- saturn_qstat(session, "-a")
cat(rawToChar(r$stdout))
}

}
