% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_upload}
\alias{saturn_upload}
\title{Saturn Upload}
\usage{
saturn_upload(session, files, to = ".", verbose = FALSE)
}
\arguments{
\item{session}{ssh connection created with \code{\link{saturn_connect}}}

\item{files}{path to files or directory to transfer}

\item{to}{existing directory on the destination where \code{files} will be copied into}

\item{verbose}{print progress while copying files}
}
\description{
Saturn Upload
}
\examples{
\dontrun{
# recursively upload files and directories
session <- saturn_connect()
files <- c(R.home("doc"), R.home("COPYING"))
saturn_upload(session, files, "~/target")
}

}
\seealso{
\code{\link[ssh]{scp}}
}
