% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_class}
\alias{select_via_class}
\title{Select columns by class}
\usage{
select_via_class(.data, class)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{class}{the class of variable to select based on (can be negated with `-`)}
}
\value{
an object of the same type as .data
}
\description{
Select columns by class
}
\examples{
## select factor columns
select_via_class(iris, factor)

## select everything but factor columns
select_via_class(iris, -factor)

## select numeric
select_via_class(iris, numeric)
}
