% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_complete_rate}
\alias{select_via_complete_rate}
\title{Subset columns by variable completion rate}
\usage{
select_via_complete_rate(.data, complete_rate)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{complete_rate}{minimally acceptable completion rate}
}
\value{
an object of the same type as .data
}
\description{
Subset columns by variable completion rate
}
\examples{
## select columns containing less than 5\% missing
select_via_complete_rate(airquality, 0.95)
}
