% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_missing_rate}
\alias{select_via_missing_rate}
\title{Subset columns by variable missing rate}
\usage{
select_via_missing_rate(.data, rate, greater = TRUE)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy
data frame (e.g. from dbplyr or dtplyr)}

\item{rate}{missing rate}

\item{greater}{if TRUE, rate is the maximally acceptable missing rate}
}
\value{
an object of the same type as .data
}
\description{
Subset columns by variable missing rate
}
\examples{
## select columns with a maximum missing rate of 5\%
select_via_missing_rate(airquality, 0.05)
}
