% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{send_mailgun}
\alias{send_mailgun}
\title{Send an email message via the Mailgun API}
\usage{
send_mailgun(
  to = "robin@wustl.edu",
  from = "Excited User",
  subject = "Mailgun Test",
  html = "<div>Test</div>",
  text = html,
  domain_name = Sys.getenv("MAILGUN_DOMAIN_NAME"),
  api_key = Sys.getenv("MAILGUN_API_KEY"),
  ssl_verifypeer = TRUE
)
}
\arguments{
\item{to}{email address of recipient}

\item{from}{name of sender (but will come from mailgun@domain_name)}

\item{subject}{subject of message}

\item{html}{html body of message}

\item{text}{text body of message}

\item{domain_name}{Mailgun domain name (e.g., sandbox<sandbox>.mailgun.org)}

\item{api_key}{Mailgun API password}

\item{ssl_verifypeer}{whether to verify peer}
}
\value{
TRUE if successful; error otherwise
}
\description{
Send an email message via the Mailgun API
}
\examples{
\dontrun{
## Retrieve Mailgun domain name and API key from .Renviron file
send_mailgun(to = "robin@wustl.edu",
             from = "Robin",
             subject = "Robin Says Hi!",
             html = "<p>Hi Robin,</p><p>Robin</p>")
}
}
\references{
\href{https://documentation.mailgun.com/en/latest/}{Mailgun API Documentation}
}
\seealso{
\code{\link[httr]{POST}}, \code{\link[httr]{set_config}}
}
