% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjPlot.R
\name{sjPlot_tab_model}
\alias{sjPlot_tab_model}
\title{Print regression models as HTML table}
\usage{
sjPlot_tab_model(
  ...,
  show.omnibus.f = FALSE,
  show.auc = FALSE,
  show.hosmer_lemeshow = FALSE,
  show.deviance_test = FALSE,
  footnote = NULL,
  digits = 2,
  digits.p = 3,
  file = NULL,
  use.viewer = TRUE,
  print = TRUE
)
}
\arguments{
\item{...}{arguments passed on to \link[sjPlot]{tab_model}; notably one or more regression models to summarize}

\item{show.omnibus.f}{Logical, if TRUE, the omnibus F-test is computed and printed}

\item{show.auc}{Logical, if TRUE, the AUC is calculated and printed}

\item{show.hosmer_lemeshow}{Logical, if TRUE, a Hosmer-Lemeshow goodness of fit test is calculated and printed}

\item{show.deviance_test}{Logical, if TRUE, a deviance-test comparing to the null model is calculated and printed}

\item{footnote}{optional footnote for table}

\item{digits}{Amount of decimals for estimates}

\item{digits.p}{Amount of decimals for p-values}

\item{file}{Destination file, if the output should be saved as file.
If \code{NULL} (default), the output will be saved as temporary file and
openend either in the IDE's viewer pane or the default web browser.}

\item{use.viewer}{Logical, if \code{TRUE}, the HTML table is shown in the IDE's
viewer pane. If \code{FALSE} or no viewer available, the HTML table is
opened in a web browser.}

\item{print}{Logical, if TRUE in non-interactive mode, the html table code is inserted into the document}
}
\value{
Invisibly returns
         \itemize{
           \item the web page style sheet (\code{page.style}),
           \item the web page content (\code{page.content}),
           \item the complete html-output (\code{page.complete}) and
           \item the html-table with inline-css for use with knitr (\code{knitr})
           }
           for further use.
}
\description{
\code{tab_model()} creates HTML tables from regression models.
}
\details{
\subsection{Standardized Estimates}{
  Default standardization is done by completely refitting the model on the
  standardized data. Hence, this approach is equal to standardizing the
  variables before fitting the model, which is particularly recommended for
  complex models that include interactions or transformations (e.g., polynomial
  or spline terms). When \code{show.std = "std2"}, standardization of estimates
  follows \href{http://www.stat.columbia.edu/~gelman/research/published/standardizing7.pdf}{Gelman's (2008)}
  suggestion, rescaling the estimates by dividing them by two standard deviations
  instead of just one. Resulting coefficients are then directly comparable for
  untransformed binary predictors. For backward compatibility reasons,
  \code{show.std} also may be a logical value; if \code{TRUE}, normal standardized
  estimates are printed (same effect as \code{show.std = "std"}). Use
  \code{show.std = NULL} (default) or \code{show.std = FALSE}, if no standardization
  is required.
}
\subsection{How do I use \code{CSS}-argument?}{
   With the \code{CSS}-argument, the visual appearance of the tables
   can be modified. To get an overview of all style-sheet-classnames
   that are used in this function, see return value \code{page.style} for details.
   Arguments for this list have following syntax:
   \enumerate{
     \item the class-names with \code{"css."}-prefix as argument name and
     \item each style-definition must end with a semicolon
   }
   You can add style information to the default styles by using a + (plus-sign) as
   initial character for the argument attributes. Examples:
   \itemize{
     \item \code{css.table = 'border:2px solid red;'} for a solid 2-pixel table border in red.
     \item \code{css.summary = 'font-weight:bold;'} for a bold fontweight in the summary row.
     \item \code{css.lasttablerow = 'border-bottom: 1px dotted blue;'} for a blue dotted border of the last table row.
     \item \code{css.colnames = '+color:green'} to add green color formatting to column names.
     \item \code{css.arc = 'color:blue;'} for a blue text color each 2nd row.
     \item \code{css.caption = '+color:red;'} to add red font-color to the default table caption style.
   }
}
}
\references{
Hosmer, David W.; Lemeshow, Stanley (2013). Applied Logistic Regression. New York: Wiley.
}
