% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sjPlot.R
\name{tab_anova}
\alias{tab_anova}
\title{Print a fitted model as HTML ANOVA table}
\usage{
tab_anova(
  mdl,
  title = NULL,
  pred.labels = NULL,
  full_width = NULL,
  digits = getOption("digits"),
  scientific = FALSE,
  ...
)
}
\arguments{
\item{mdl}{a model returned by a model fitting function (e.g., lm or glm)}

\item{title}{character, will be used as table caption.}

\item{pred.labels}{character vector with labels of predictor varaibles}

\item{full_width}{logical controlling whether the html table should have 100\% width}

\item{digits}{amount of decimals for numbers}

\item{scientific}{logical specifying whether to encode numbers in scientific format}
}
\value{
a character vector of the table source code
}
\description{
Print a fitted model as HTML ANOVA table
}
\examples{
m <- lm(mpg ~ cyl, mtcars)
tab_anova(m)

}
\seealso{
\link[sjPlot]{tab_model}, \link[stats]{anova}, \link[base]{format}, \link[kableExtra]{kbl}, \link[kableExtra]{kable_classic}
}
