% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Generate an HTML table of descriptive statistics}
\usage{
table1(
  x,
  data,
  file,
  css = system.file("table1_defaults_1.0", "table1_defaults.css", package = "table1"),
  ...
)
}
\arguments{
\item{x}{An object, typically a \code{formula} or list of \code{data.frame}s (see Details).}

\item{data}{For the formula interface, a \code{data.frame} from which the
variables in \code{x} should be taken.}

\item{file}{optional file location to write table to}

\item{css}{cascading style sheet to embed}

\item{...}{Further arguments, passed to \code{render}.}
}
\value{
a washu_table1 object
}
\description{
Generate an HTML table of descriptive statistics
}
\examples{
\dontrun{
table1(~., mtcars, "Table 1. Descriptives.html")
}
}
