% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tempfile_html}
\alias{tempfile_html}
\title{Create Names for Temporary HTML File}
\usage{
tempfile_html(pattern = "file", tmpdir = tempdir())
}
\description{
Create Names for Temporary HTML File
}
\section{Functions}{
\itemize{
\item \code{tempfile_html()}: tempfile_html returns a vector of character strings which can be used as names for temporary html files.

}}
