% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tempfile_html}
\alias{tempfile_html}
\title{Create Names for Temporary HTML File}
\usage{
tempfile_html(pattern = "file", tmpdir = tempdir())
}
\arguments{
\item{pattern}{a non-empty character vector giving the initial part of the
name.}

\item{tmpdir}{a non-empty character vector giving the directory name}
}
\description{
Create Names for Temporary HTML File
}
