% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{tex_pigletsignature}
\alias{tex_pigletsignature}
\title{TeX new pigletsignature command}
\usage{
tex_pigletsignature()
}
\value{
character vector of TeX code
}
\description{
TeX new pigletsignature command
}
