% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{tex_runtime_in_header}
\alias{tex_runtime_in_header}
\title{TeX runtime header additions}
\usage{
tex_runtime_in_header(x)
}
\arguments{
\item{x}{character vector of TeX code}
}
\value{
path to temporary file containing the TeX code
}
\description{
TeX runtime header additions
}
