% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{tex_setmainfont}
\alias{tex_setmainfont}
\title{TeX Setmainfont}
\usage{
tex_setmainfont(font = "LibreBaskerville", ext = "ttf")
}
\arguments{
\item{font}{desired font}

\item{ext}{font extention}
}
\value{
character vector of TeX code
}
\description{
TeX Setmainfont
}
