% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{tidy_sub}
\alias{tidy_sub}
\title{Tidy Sub
Utility function to make sub() work better with pipes}
\usage{
tidy_sub(x, pattern, replacement, fixed = TRUE, ...)
}
\arguments{
\item{x}{a character vector where matches are sought}

\item{pattern}{character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector.}

\item{replacement}{a replacement for matched pattern in sub and gsub.}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is.}

\item{...}{additional parameters passed to sub}
}
\description{
Tidy Sub
Utility function to make sub() work better with pipes
}
