% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{valid_project_name}
\alias{valid_project_name}
\title{Valid project name
Check to see if project name follow YYYY-MM-DD-N format}
\usage{
valid_project_name(project_name, version = FALSE)
}
\arguments{
\item{project_name}{proposed project name to check}

\item{version}{flag to check for version number}
}
\description{
Valid project name
Check to see if project name follow YYYY-MM-DD-N format
}
