% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{webplotdigitizer}
\alias{webplotdigitizer}
\title{WebPlotDigitizer
Open the WebPlotDigitizer web app in the default browser.}
\usage{
webplotdigitizer()
}
\description{
WebPlotDigitizer
Open the WebPlotDigitizer web app in the default browser.
}
\details{
It is often necessary to reverse engineer images of data visualizations to extract the underlying numerical data. WebPlotDigitizer is a semi-automated tool that makes this process extremely easy:
}
