% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webpower.R
\name{WebPower_bind_webpower}
\alias{WebPower_bind_webpower}
\title{WebPower Bind WebPower Objects}
\usage{
WebPower_bind_webpower(x, y)
}
\arguments{
\item{x}{first WebPower object}

\item{y}{second WebPower object}
}
\value{
a combined WebPower object
}
\description{
WebPower Bind WebPower Objects
}
\examples{
\dontrun{
WebPower::wp.rmanova(
  n = 40,
  ng = 2,
  nm = 2,
  f = NULL,
  nscor = 1,
  alpha = 0.05,
  power = 0.80,
  type = 2
) -> x

WebPower::wp.rmanova(
  n = 200,
  ng = 2,
  nm = 2,
  f = NULL,
  nscor = 1,
  alpha = 0.05,
  power = 0.80,
  type = 2
) -> y

WebPower_bind_webpower(x, y)
}
}
