% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{win_attrib}
\alias{win_attrib}
\title{Windows attrib command}
\usage{
win_attrib(x, a = TRUE, h = TRUE, s = TRUE)
}
\arguments{
\item{x}{file for which to set attributes}

\item{a}{set archive flag}

\item{h}{set hide flag}

\item{s}{set system flag}
}
\description{
Windows attrib command
}
\examples{
\dontrun{
win_attrib("hide me.lnk")
}
}
