% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_web_link_file}
\alias{write_web_link_file}
\title{Write web shortcut file}
\usage{
write_web_link_file(con, url, title = "", ...)
}
\arguments{
\item{con}{A \link[base]{connections} object or character string}

\item{url}{URL of location}

\item{title}{title of link page}

\item{...}{additional arguments passed to \link[base]{writeLines}}
}
\description{
Write web shortcut file
}
