% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{wu_colors}
\alias{wu_colors}
\title{WU colors}
\format{
An object of class \code{list} of length 16.
}
\usage{
wu_colors
}
\description{
WashU Color Palettes
}
\seealso{
\url{https://publicaffairs.wustl.edu/assets/color-palettes/}
}
\keyword{datasets}
