% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio.R
\name{wu_consult_project}
\alias{wu_consult_project}
\title{Consult Project}
\usage{
wu_consult_project(path, principal_investigator = "Dr. Winnie Pooh", ...)
}
\arguments{
\item{path}{path to new project}

\item{principal_investigator}{name of the principal investigator}

\item{...}{parameters passed from the new project dialog}
}
\description{
Consult Project
}
