% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{wu_directory}
\alias{wu_directory}
\title{WU directory search}
\usage{
wu_directory(search_name = "", email = "", phone = "")
}
\arguments{
\item{search_name}{name query}

\item{email}{email query}

\item{phone}{phone query}
}
\value{
data frame of results
}
\description{
Search the WashU directory
}
\note{
Will return an empty data frame if there are none or too many results
}
\examples{
wu_directory("schuelke")
}
\seealso{
\url{https://wustl.edu/directory/}
}
