% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{wu_emph}
\alias{wu_emph}
\title{WU emphasis}
\usage{
wu_emph(x, color = wu_colors$red, bold = TRUE)
}
\arguments{
\item{x}{text to span}

\item{color}{color to use}

\item{bold}{= TRUE}
}
\value{
html span element with requested inline css attributes
}
\description{
WU emphasis
}
\examples{
wu_emph("I will be bold red text.")
}
