% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{wu_ldap_query_default}
\alias{wu_ldap_query_default}
\title{WU LDAP query}
\usage{
wu_ldap_query_default(
  hostname = "accounts.ad.wustl.edu",
  base_dn = "OU=Current,OU=People,DC=accounts,DC=ad,DC=wustl,DC=edu",
  attributes = "",
  scope = "sub",
  filter = "(objectClass=*)",
  user,
  pass
)
}
\arguments{
\item{hostname}{name (or ip address) of the LDAP server}

\item{base_dn}{distinguished name (DN) of an entry in the directory. This DN identifies the entry that is the starting point of the search. If no base DN is specified, the search starts at the root of the directory tree.}

\item{attributes}{The attributes to be returned. To specify more than one attribute, use commas to separate the attributes (for example, "cn,mail,telephoneNumber"). If no attributes are specified in the URL, all attributes are returned.}

\item{scope}{The scope of the search, which can be one of these values:
\itemize{
  \item{base}{retrieves information about the distinguished name (base_dn) specified in the URL only.}
  \item{one}{retrieves information about entries one level below the distinguished name (base_dn) specified in the URL. The base entry is not included in this scope.}
  \item{sub}{retrieves information about entries at all levels below the distinguished name (base_dn) specified in the URL. The base entry is included in this scope.}
}}

\item{filter}{Search filter to apply to entries within the specified scope of the search.}

\item{user}{username}

\item{pass}{password}
}
\value{
curl response object
}
\description{
Default values for WU LDAP queries
}
\seealso{
\code{\link{ldap_query}}
}
