% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{wu_ldap_query_mail}
\alias{wu_ldap_query_mail}
\title{WU LDAP query for mail}
\usage{
wu_ldap_query_mail(mail, user, pass)
}
\arguments{
\item{mail}{mail to search for}

\item{user}{searcher username}

\item{pass}{searcher password}
}
\value{
tibble of all attributes for given mail
}
\description{
Search WU LDAP for a given mail
}
