% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{wu_ldap_query_sAMAccountName}
\alias{wu_ldap_query_sAMAccountName}
\title{WU LDAP query for sAMAccountName}
\usage{
wu_ldap_query_sAMAccountName(sAMAccountName, user, pass)
}
\arguments{
\item{sAMAccountName}{sAMAccountName to search for}

\item{user}{searcher username}

\item{pass}{searcher password}
}
\value{
tibble of all attributes for given sAMAccountName
}
\description{
Search WU LDAP for a given sAMAccountName
}
