% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldap.R
\name{wu_ldap_query_userPrincipalName}
\alias{wu_ldap_query_userPrincipalName}
\title{WU LDAP query for userPrincipalName}
\usage{
wu_ldap_query_userPrincipalName(userPrincipalName, user, pass)
}
\arguments{
\item{userPrincipalName}{userPrincipalName to search for}

\item{user}{searcher username}

\item{pass}{searcher password}
}
\value{
tibble of all attributes for given userPrincipalName
}
\description{
Search WU LDAP for a given userPrincipalName
}
