% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{wu_render_estimate}
\alias{wu_render_estimate}
\title{Render estimate}
\usage{
wu_render_estimate(
  ref,
  date,
  description,
  to_name,
  to_title,
  to_campus_box,
  to_email,
  from_name,
  from_title,
  from_campus_box,
  from_email,
  data,
  input = "Estimate.Rmd",
  keep_input = FALSE,
  ...
)
}
\arguments{
\item{ref, date, description}{estimate reference, date, and description}

\item{to_name, to_title, to_campus_box, to_email}{recipient attributes}

\item{from_name, from_title, from_campus_box, from_email}{sender attributes}

\item{data}{path to estimate data object containing service, hours, and rate
columns}

\item{input}{The input file to be rendered. This can be an R script (.R),
an R Markdown document (.Rmd), or a plain markdown document.}

\item{keep_input}{keep the input document}

\item{...}{parameters passed to \code{\link[rmarkdown]{render}}}
}
\description{
Render estimate
}
\examples{
\dontrun{
## some yaml may require quotes
wu_render_estimate(
  paste0(Sys.Date(), "-1"),
  Sys.Date(),
  "The Woozle Effect",
  "Christopher Robin",
  "Billy Moon",
  "1920",
  "robin@wustl.edu",
  "Matthew Schuelke, PhD",
  "Research Statistician",
  "8067",
  "schuelke@wustl.edu",
  "dat/est.csv"
)
}
}
