% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimr.R
\name{wu_skim_with}
\alias{wu_skim_with}
\title{Set or add the summary functions for a particular type of data}
\usage{
wu_skim_with(
  fct_n_levels = NULL,
  fct_percents = TRUE,
  fct_digits = 1,
  fct_delimiter = c(", ", "<br>")
)
}
\arguments{
\item{fct_n_levels}{number of factor levels to print sorted in descending
frequency}

\item{fct_percents}{flag to print percents per level}

\item{fct_digits}{number of significant digits to print for factor summaries}

\item{fct_delimiter}{delimiter to use between factor levels}
}
\value{
A new skim() function. This is callable.
}
\description{
Build a custom skim function that adds percents to factor
summaries
}
\note{
NAs are ignored when computing percents.
}
\examples{
wu_skim <- wu_skim_with()
wu_skim(factor(mtcars$cyl))
}
