% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.R
\name{xl_column_number}
\alias{xl_column_number}
\title{Excel column number}
\usage{
xl_column_number(column_reference)
}
\arguments{
\item{column_reference}{excel column reference (e.g., "A" or "AC")}
}
\value{
column number index
}
\description{
Excel column number
}
\examples{
xl_column_number("A")
}
