% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R, R/zotero.R
\name{bib_add_pkg_zotero}
\alias{bib_add_pkg_zotero}
\alias{zt_add_pkg}
\title{Zotero Add Package}
\usage{
bib_add_pkg_zotero(package, collection_name, ...)

zt_add_pkg(package, collection_name, ...)
}
\arguments{
\item{package}{a character string with the name of a single package}

\item{collection_name}{name of collection to add citation to}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\description{
Zotero Add Package
}
