% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliography.R, R/zotero.R
\name{bib_use_zotero}
\alias{bib_use_zotero}
\alias{zt_create_collection}
\title{Zotero create collection}
\usage{
bib_use_zotero(
  collection_name = bib_zotero_collection(),
  parent_collection_key = NULL,
  ...
)

zt_create_collection(collection_name, parent_collection_key = NULL, ...)
}
\arguments{
\item{collection_name}{name of the collection}

\item{parent_collection_key}{collection key of the parent collections}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\description{
Zotero create collection
}
