% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_get_users_collections}
\alias{zt_get_users_collections}
\title{Zotero GET /users/<api_user>/collections}
\usage{
zt_get_users_collections(path = "", ...)
}
\arguments{
\item{path}{optional additional path}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
\link[httr]{response} object
}
\description{
Zotero GET /users/<api_user>/collections
}
