% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_get_users_collections_items_top}
\alias{zt_get_users_collections_items_top}
\title{Zotero GET /users/<api_user>/collections/<collection_key>/items/top}
\usage{
zt_get_users_collections_items_top(collection_key, ...)
}
\arguments{
\item{collection_key}{Zotero collection key}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
\link[httr]{response} object
}
\description{
Zotero GET /users/<api_user>/collections/<collection_key>/items/top
}
