% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_lookup_user_collection_key}
\alias{zt_lookup_user_collection_key}
\title{Zotero lookup collection key by collection name}
\usage{
zt_lookup_user_collection_key(collection_name, ...)
}
\arguments{
\item{collection_name}{name of the collection}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
collection key
}
\description{
Zotero lookup collection key by collection name
}
