% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_lookup_user_collections_name_key_pairs}
\alias{zt_lookup_user_collections_name_key_pairs}
\title{Zotero lookup name-key pairs for a user's collections}
\usage{
zt_lookup_user_collections_name_key_pairs(...)
}
\arguments{
\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
named character vector
}
\description{
Zotero lookup name-key pairs for a user's collections
}
