% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_post}
\alias{zt_post}
\title{Zotero POST}
\usage{
zt_post(
  url = "https://api.zotero.org",
  api_key = Sys.getenv("ZOTERO_API_KEY"),
  ...
)
}
\arguments{
\item{url}{the url of the page to retrieve}

\item{api_key}{Zotero API key}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
\link[httr]{response} object
}
\description{
Zotero POST
}
