% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zotero.R
\name{zt_post_users}
\alias{zt_post_users}
\title{Zotero POST /users/<api_user>}
\usage{
zt_post_users(path = "", api_user = Sys.getenv("ZOTERO_API_USER"), ...)
}
\arguments{
\item{path}{optional path}

\item{api_user}{Zotero API user}

\item{...}{Further named parameters, such as \code{query}, \code{path}, etc,
passed on to \code{\link[httr:modify_url]{modify_url()}}. Unnamed parameters will be combined
with \code{\link[httr:config]{config()}}.}
}
\value{
\link[httr]{response} object
}
\description{
Zotero POST /users/<api_user>
}
