% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broom.R
\name{broom_tidy.glmerMod}
\alias{broom_tidy.glmerMod}
\title{Tidy a glmerMod object}
\usage{
broom_tidy.glmerMod(
  x,
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE,
  ...
)
}
\arguments{
\item{x}{A glmerMod object returned from \code{\link[lme4:glmer]{glmer()}}}

\item{conf.int}{Logical indicating whether or not to include a confidence
interval in the tidied output. Defaults to FALSE.}

\item{conf.level}{The confidence level to use for the confidence interval if
conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to
 0.95, which corresponds to a 95 percent confidence interval.}

\item{exponentiate}{Logical indicating whether or not to exponentiate the the
coefficient estimates. This is typical for logistic and multinomial
regressions, but a bad idea if there is no log or logit link. Defaults to FALSE.}

\item{...}{Additional arguments. Not used. Needed to match generic signature
only. Cautionary note: Misspelled arguments will be absorbed in ..., where
they will be ignored. If the misspelled argument has a default value, the
default value will be used.}
}
\value{
A \link[tibble]{tibble} with information about fixed model estimates.
}
\description{
Tidy a glmerMod object
}
\examples{
m <- lme4::glmer(vs ~ mpg + (1 | cyl), mtcars, binomial)
broom_tidy.glmerMod(m)
}
