% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check}
\alias{check}
\title{Build and check a package, cleaning up automatically on success.}
\usage{
check(document = FALSE, env_vars = c(`_R_CHECK_SYSTEM_CLOCK_` = FALSE), ...)
}
\arguments{
\item{document}{By default (\code{NULL}) will document if your installed
roxygen2 version matches the version declared in the \code{DESCRIPTION}
file. Use \code{TRUE} or \code{FALSE} to override the default.}

\item{env_vars}{Environment variables set during \verb{R CMD check}}

\item{...}{additional parameters passed to \code{\link[devtools]{check}}}
}
\description{
Build and check a package, cleaning up automatically on success.
}
\note{
Adds environment flag to skip time check as per
\href{https://stat.ethz.ch/pipermail/r-package-devel/2019q1/003577.html}{r-package-devel}.
}
