% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consult_database.R
\name{db_search_consults}
\alias{db_search_consults}
\title{Run shiny app to search consult database}
\usage{
db_search_consults(db = Sys.getenv("WU_CONSULT_DB"))
}
\arguments{
\item{db}{location of the database}
}
\description{
Run shiny app to search consult database
}
