% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance_test.R
\name{print.dtest}
\alias{print.dtest}
\title{Print Objects of Type dtest}
\usage{
\method{print}{dtest}(x, format = c("tibble", "text"), digits = 4, ...)
}
\arguments{
\item{x}{an object of class "dtype", a result of a call to \code{\link[deviance.test]{washu::deviance.test()}}}

\item{format}{format of the output}

\item{digits}{number of decimal digits to display}

\item{...}{additional parameters passed to other functions}
}
\value{
the argument x, invisibly
}
\description{
Print Objects of Type dtest
}
