% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saturn.R
\name{saturn_job_init}
\alias{saturn_job_init}
\title{Saturn Job Initialize}
\usage{
saturn_job_init(
  job_name,
  spath = ".",
  nodes = 1,
  ppn = 4,
  user = Sys.getenv("WUSTL_KEY_USER"),
  property = c("", "R", "SAS", "MPlus", "EL7"),
  walltime = "00:00:00",
  queue = c("fast", "quick", "batch", "interactive"),
  M,
  m = "abe",
  edit_script = TRUE
)
}
\arguments{
\item{job_name}{job name}

\item{spath}{directory on the destination where files will be copied into}

\item{nodes}{number of nodes}

\item{ppn}{processors per node}

\item{user}{user account}

\item{property}{requested node property}

\item{walltime}{total time requested for the job}

\item{queue}{queue name (note, this will be auto redirected depending on resource request)}

\item{M}{email address to send notifications}

\item{m}{email notifications: (a) abort, (b) begin, (e) end}

\item{edit_script}{open script for editing}
}
\description{
Saturn Job Initialize
}
\examples{
\dontrun{
job_name <- "my_job"
saturn_job_init(job_name)
}
}
