% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_class}
\alias{select_via_class}
\title{Select columns by class}
\usage{
select_via_class(.data, class, keep = TRUE)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{class}{variable class to select on}

\item{keep}{if TRUE, keep variables of type class, otherwise drop}
}
\value{
an object of the same type as .data
}
\description{
Select columns by class
}
\examples{
## select factor columns
select_via_class(iris, "factor")

## select everything but factor columns
select_via_class(iris, "factor", FALSE)

## select numeric
select_via_class(iris, "numeric")
}
