% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_via_complete_rate}
\alias{select_via_complete_rate}
\title{Subset columns by variable completion rate}
\usage{
select_via_complete_rate(.data, rate, greater = TRUE)
}
\arguments{
\item{.data}{a data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr)}

\item{rate}{completion rate}

\item{greater}{if TRUE, rate is the minimally acceptable completion rate}
}
\value{
an object of the same type as .data
}
\description{
Subset columns by variable completion rate
}
\examples{
## select columns with at least a 95\% completion rate
select_via_complete_rate(airquality, 0.95)
}
