% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{wu_colors_plot}
\alias{wu_colors_plot}
\title{WU Colors Plot}
\usage{
wu_colors_plot()
}
\description{
WashU Color Palettes Plot
}
\examples{
wu_colors_plot()
}
\seealso{
\url{https://publicaffairs.wustl.edu/assets/color-palettes/}
}
