% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{wu_render_letter_of_support}
\alias{wu_render_letter_of_support}
\title{Render letter of support}
\usage{
wu_render_letter_of_support(
  from_name,
  from_title,
  from_department_type,
  from_department_name,
  from_department_url,
  from_campus_box,
  from_phone,
  from_email,
  to_name,
  to_address,
  date,
  salutation,
  closing,
  body,
  signature = "",
  input = "LOS.Rmd",
  keep_input = FALSE,
  ...
)
}
\arguments{
\item{from_name, from_title, from_department_type, from_department_name, from_department_url, from_campus_box, from_phone, from_email}{attributes of sender}

\item{to_name, to_address}{attributes of recipient}

\item{date, salutation, closing}{letter customization}

\item{body}{body of letter}

\item{signature}{path to signature file (empty string for no signature)}

\item{input}{The input file to be rendered. This can be an R script (.R),
an R Markdown document (.Rmd), or a plain markdown document.}

\item{keep_input}{keep the input document}

\item{...}{parameters passed to \code{\link[rmarkdown]{render}}}
}
\description{
Render letter of support
}
\examples{
\dontrun{
## some yaml may require quotes, e.g.:
## - salutation because of the colon
## - signature if spaces in path
wu_render_letter_of_support(
  "Matthew J. Schuelke, PhD",
  "Research Statistician",
  "Division",
  "Biostatistics",
  "https://biostatistics.wustl.edu",
  "8067",
  "+1 314 362-0000",
  "schuelke@wustl.edu",
  "Christopher Robin",
  c("11 Mallord Street", "LONDON", "SW3 6DT", "UK"),
  Sys.Date(),
  "\"To Whom It May Concern:\"",
  "Sincerely,",
  letter_of_support_body("The Woozle Effect"),
  paste0("\"", Sys.getenv("WU_SIGNATURE"), "\"")
)
}
}
